# Faster gcd for double-limb-case.