--- a/include/boost/multiprecision/cpp_int/misc.hpp
+++ b/include/boost/multiprecision/cpp_int/misc.hpp
@@ -337,17 +337,21 @@ inline BOOST_MP_CXX14_CONSTEXPR typename enable_if_c<is_integral<U>::value>::typ
 
 template <unsigned MinBits1, unsigned MaxBits1, cpp_integer_type SignType1, cpp_int_check_type Checked1, class Allocator1, class Integer>
 inline BOOST_MP_CXX14_CONSTEXPR typename enable_if_c<is_unsigned<Integer>::value && !is_trivial_cpp_int<cpp_int_backend<MinBits1, MaxBits1, SignType1, Checked1, Allocator1> >::value, Integer>::type
-eval_integer_modulus(const cpp_int_backend<MinBits1, MaxBits1, SignType1, Checked1, Allocator1>& x, Integer val)
+eval_integer_modulus(const cpp_int_backend<MinBits1, MaxBits1, SignType1, Checked1, Allocator1>& a, Integer mod)
 {
-   if ((sizeof(Integer) <= sizeof(limb_type)) || (val <= (std::numeric_limits<limb_type>::max)()))
+   if ((sizeof(Integer) <= sizeof(limb_type)) || (mod <= (std::numeric_limits<limb_type>::max)()))
    {
-      cpp_int_backend<MinBits1, MaxBits1, SignType1, Checked1, Allocator1> d;
-      divide_unsigned_helper(static_cast<cpp_int_backend<MinBits1, MaxBits1, SignType1, Checked1, Allocator1>*>(0), x, static_cast<limb_type>(val), d);
-      return d.limbs()[0];
+      const int              n = a.size();
+      const double_limb_type two_n_mod = static_cast<limb_type>(1u) + (~static_cast<limb_type>(0u) - mod) % mod;
+      limb_type              res = a.limbs()[n - 1] % mod;
+
+      for (int i = n - 2; i >= 0; --i)
+         res = (res * two_n_mod + a.limbs()[i]) % mod;
+      return res;
    }
    else
    {
-      return default_ops::eval_integer_modulus(x, val);
+      return default_ops::eval_integer_modulus(a, mod);
    }
 }
 
@@ -368,30 +372,6 @@ inline BOOST_MP_CXX14_CONSTEXPR limb_type integer_gcd_reduce(limb_type u, limb_t
          break;
       v -= u;
       v >>= boost::multiprecision::detail::find_lsb(v);
-   } while (true);
-   return u;
-}
-
-inline BOOST_MP_CXX14_CONSTEXPR double_limb_type integer_gcd_reduce(double_limb_type u, double_limb_type v)
-{
-   do
-   {
-      if (u > v)
-         std_constexpr::swap(u, v);
-      if (u == v)
-         break;
-      if (v <= ~static_cast<limb_type>(0))
-      {
-         u = integer_gcd_reduce(static_cast<limb_type>(v), static_cast<limb_type>(u));
-         break;
-      }
-      v -= u;
-#ifdef __MSVC_RUNTIME_CHECKS
-      while ((v & 1u) == 0)
-#else
-      while ((static_cast<unsigned>(v) & 1u) == 0)
-#endif
-         v >>= 1;
    } while (true);
    return u;
 }
@@ -417,6 +397,30 @@ BOOST_MP_FORCEINLINE BOOST_MP_CXX14_CONSTEXPR limb_type eval_gcd(limb_type u, li
 #endif
 }
 
+inline BOOST_MP_CXX14_CONSTEXPR double_limb_type integer_gcd_reduce(double_limb_type u, double_limb_type v)
+{
+   do
+   {
+      if (u > v)
+         std_constexpr::swap(u, v);
+      if (u == v)
+         break;
+      if (v <= ~static_cast<limb_type>(0))
+      {
+         u = eval_gcd(static_cast<limb_type>(v), static_cast<limb_type>(u));
+         break;
+      }
+      v -= u;
+#ifdef __MSVC_RUNTIME_CHECKS
+      while ((v & 1u) == 0)
+#else
+      while ((static_cast<unsigned>(v) & 1u) == 0)
+#endif
+         v >>= 1;
+   } while (true);
+   return u;
+}
+
 template <unsigned MinBits1, unsigned MaxBits1, cpp_integer_type SignType1, cpp_int_check_type Checked1, class Allocator1>
 inline BOOST_MP_CXX14_CONSTEXPR typename enable_if_c<!is_trivial_cpp_int<cpp_int_backend<MinBits1, MaxBits1, SignType1, Checked1, Allocator1> >::value>::type
 eval_gcd(
@@ -434,6 +438,26 @@ eval_gcd(
    eval_modulus(result, a, b);
    limb_type& res = *result.limbs();
    res            = eval_gcd(res, b);
+   result.sign(false);
+}
+
+template <unsigned MinBits1, unsigned MaxBits1, cpp_integer_type SignType1, cpp_int_check_type Checked1, class Allocator1>
+inline BOOST_MP_CXX14_CONSTEXPR typename enable_if_c<!is_trivial_cpp_int<cpp_int_backend<MinBits1, MaxBits1, SignType1, Checked1, Allocator1> >::value>::type
+eval_gcd(
+    cpp_int_backend<MinBits1, MaxBits1, SignType1, Checked1, Allocator1>&       result,
+    const cpp_int_backend<MinBits1, MaxBits1, SignType1, Checked1, Allocator1>& a,
+    double_limb_type                                                            b)
+{
+   int s = eval_get_sign(a);
+   if (!b || !s)
+   {
+      result = a;
+      *result.limbs() |= b;
+      return;
+   }
+   double_limb_type res = eval_integer_modulus(a, b);
+   res            = eval_gcd(res, b);
+   result = res;
    result.sign(false);
 }
 
@@ -524,7 +548,7 @@ eval_gcd(
       if (v.size() <= 2)
       {
          if (v.size() == 1)
-            u = integer_gcd_reduce(*v.limbs(), *u.limbs());
+            u = eval_gcd(*v.limbs(), *u.limbs());
          else
          {
             double_limb_type i = v.limbs()[0] | (static_cast<double_limb_type>(v.limbs()[1]) << sizeof(limb_type) * CHAR_BIT);