Using the performance test from @madhur4127 I see:

```
Before:
BM<cpp_int>/400        8202 ns         7952 ns        74667
BM<cpp_int>/600       14937 ns        14579 ns        40727
BM<cpp_int>/800       21660 ns        21449 ns        29867
BM<cpp_int>/1000      25239 ns        25112 ns        28000
BM<cpp_int>/1200      34091 ns        34424 ns        21333
BM<cpp_int>/1400      43222 ns        43493 ns        15448
BM<cpp_int>/1600      53205 ns        53013 ns        11200
BM<cpp_int>/1800      62238 ns        62500 ns        10000
BM<cpp_int>/2000      71333 ns        69754 ns         8960

after:
BM<cpp_int>/400        2251 ns         2176 ns       373333
BM<cpp_int>/600        2659 ns         2651 ns       224000
BM<cpp_int>/800        3826 ns         3770 ns       194783
BM<cpp_int>/1000       4351 ns         4297 ns       160000
BM<cpp_int>/1200       5064 ns         4604 ns       112000
BM<cpp_int>/1400       6112 ns         6138 ns       112000
BM<cpp_int>/1600       8036 ns         7952 ns       112000
BM<cpp_int>/1800       9050 ns         8894 ns        89600
BM<cpp_int>/2000      10376 ns        10463 ns        89600
```

With msvc.