@jzmaddock 2020-04-14
With gcc-9 on Ubuntu I see:

```
Before:
BM<cpp_int>/400        4096 ns       4096 ns     170201
BM<cpp_int>/600        6990 ns       6988 ns      97366
BM<cpp_int>/800       10709 ns      10709 ns      67967
BM<cpp_int>/1000      14156 ns      14155 ns      49726
BM<cpp_int>/1200      18216 ns      18214 ns      38690
BM<cpp_int>/1400      23105 ns      23102 ns      30572
BM<cpp_int>/1600      29206 ns      29205 ns      24718
BM<cpp_int>/1800      34244 ns      34242 ns      20436
BM<cpp_int>/2000      39595 ns      39592 ns      17476

After:
BM<cpp_int>/400         729 ns        728 ns     932376
BM<cpp_int>/600         934 ns        934 ns     748705
BM<cpp_int>/800        1143 ns       1143 ns     610194
BM<cpp_int>/1000       1405 ns       1390 ns     497985
BM<cpp_int>/1200       1648 ns       1648 ns     389981
BM<cpp_int>/1400       1971 ns       1971 ns     357603
BM<cpp_int>/1600       2270 ns       2270 ns     303139
BM<cpp_int>/1800       2594 ns       2593 ns     274015
BM<cpp_int>/2000       2895 ns       2895 ns     241997
```

Which strangely looks even better than the msvc case, I was expecting this to be the other way around, can anyone please run the test case below with and without this PR?  Thanks!

```
#include <iostream>
#include <benchmark/benchmark.h>
#include <boost/multiprecision/cpp_int.hpp>
#include <boost/multiprecision/gmp.hpp>
#include <boost/random.hpp>
#include <cmath>

using namespace boost::multiprecision;
using namespace boost::random;

template <typename T>
static void BM(benchmark::State& state)
{
   mt19937_64                  mt;
   int                         bits = state.range(0);
   uniform_int_distribution<T> ui((T(1) << (bits - 1) + 1), T(1) << bits);
   T                           x = ui(mt), z;
   unsigned __int128           y = 0x12abcdef12abcdef;
   y <<= 64;
   y |= 0x5765fde4533;
   for (auto _ : state)
   {
	  ++x;
	  ++y;
	  benchmark::DoNotOptimize(z = gcd(x, y));
   }
}
int start = 4e2, end = 2e3, step = 2e2;
BENCHMARK_TEMPLATE(BM, cpp_int)->DenseRange(start, end, step);
//BENCHMARK_TEMPLATE(BM, mpz_int)->DenseRange(start, end, step);
BENCHMARK_MAIN();
```