@jzmaddock 2020-04-14
Weird, Intel is even more of a difference:

```
Before:
BM<cpp_int>/400        7827 ns       7814 ns      89159
BM<cpp_int>/600       13081 ns      13066 ns      53363
BM<cpp_int>/800       18869 ns      18858 ns      36635
BM<cpp_int>/1000      24765 ns      24749 ns      27770
BM<cpp_int>/1200      31790 ns      31790 ns      22109
BM<cpp_int>/1400      39197 ns      39191 ns      18278
BM<cpp_int>/1600      46418 ns      46412 ns      14279
BM<cpp_int>/1800      53226 ns      53226 ns      12200
BM<cpp_int>/2000      60423 ns      60317 ns      11374

after:
BM<cpp_int>/400         743 ns        743 ns     906386
BM<cpp_int>/600         931 ns        931 ns     737220
BM<cpp_int>/800        1120 ns       1120 ns     613565
BM<cpp_int>/1000       1401 ns       1384 ns     516753
BM<cpp_int>/1200       1704 ns       1681 ns     415419
BM<cpp_int>/1400       2053 ns       2048 ns     338719
BM<cpp_int>/1600       2704 ns       2700 ns     287685
BM<cpp_int>/1800       2671 ns       2671 ns     245763
BM<cpp_int>/2000       3089 ns       3067 ns     218096
```

Which is a 20x performance improvement in the best case.... I suspect I'm missing something.