@madhur4127 2020-04-14
before: current develop
after: this branch 

```
// gcc version 9.3.0 (Arch Linux 9.3.0-1) with -O2

// before:
Benchmark                 Time             CPU   Iterations
-----------------------------------------------------------
BM<cpp_int>/400        8514 ns         8389 ns        84870
BM<cpp_int>/600       14329 ns        14146 ns        48718
BM<cpp_int>/800       21341 ns        21021 ns        33786
BM<cpp_int>/1000      29499 ns        29078 ns        24297
BM<cpp_int>/1200      37670 ns        37153 ns        18513
BM<cpp_int>/1400      48980 ns        47484 ns        14916
BM<cpp_int>/1600      58085 ns        57331 ns        11721
BM<cpp_int>/1800      71424 ns        70409 ns         9726
BM<cpp_int>/2000      86670 ns        85376 ns         8525

// after:
BM<cpp_int>/400        1332 ns         1319 ns       476268
BM<cpp_int>/600        1759 ns         1743 ns       406714
BM<cpp_int>/800        2219 ns         2201 ns       320167
BM<cpp_int>/1000       2733 ns         2702 ns       258713
BM<cpp_int>/1200       3571 ns         3514 ns       201295
BM<cpp_int>/1400       4096 ns         4052 ns       158772
BM<cpp_int>/1600       4812 ns         4770 ns       142680
BM<cpp_int>/1800       5301 ns         5257 ns       132721
BM<cpp_int>/2000       6126 ns         6064 ns       118732
```
and with clang
```
// clang version 9.0.1
//before:
Benchmark                 Time             CPU   Iterations
-----------------------------------------------------------
BM<cpp_int>/400       13124 ns        13011 ns        50181
BM<cpp_int>/600       22576 ns        22390 ns        31401
BM<cpp_int>/800       33485 ns        33184 ns        21629
BM<cpp_int>/1000      44198 ns        43873 ns        15942
BM<cpp_int>/1200      56576 ns        56140 ns        12287
BM<cpp_int>/1400      69098 ns        68536 ns         9849
BM<cpp_int>/1600      84288 ns        83680 ns         8420
BM<cpp_int>/1800      99183 ns        98370 ns         6945
BM<cpp_int>/2000     117451 ns       116302 ns         6174

// after:
Benchmark                 Time             CPU   Iterations
-----------------------------------------------------------
BM<cpp_int>/400        1427 ns         1351 ns       513522
BM<cpp_int>/600        1844 ns         1809 ns       389109
BM<cpp_int>/800        2324 ns         2285 ns       308563
BM<cpp_int>/1000       2844 ns         2760 ns       254928
BM<cpp_int>/1200       3406 ns         3325 ns       199217
BM<cpp_int>/1400       3868 ns         3812 ns       184087
BM<cpp_int>/1600       4662 ns         4564 ns       136071
BM<cpp_int>/1800       5212 ns         5129 ns       100000
BM<cpp_int>/2000       5659 ns         5588 ns       125678
```