@NAThompson 2020-04-14
@madhur4127 : Could you run it the benchmark under `perf` and see what instructions are hot?