@madhur4127 2020-04-14
@jzmaddock can you post benchmarks wrt clang? I am seeing a strange slowdown here.

@NAThompson, does this help?
Right shift generic: 37% (bitwise.hpp:578)
Subtract unsigned: 21% 
eval_gcd: 18%
eval_lsb: 11%
eval_right_shift: 11%

![image](https://user-images.githubusercontent.com/29328609/79224991-ad307c80-7e79-11ea-8ef1-9f784d8e5cd2.png)