@NAThompson 2020-04-14
@madhur4127 : This is a nice callstack, and useful information. But the asm is displayed via

```
$ perf record -g ./bench.x
$ perf report -g -M intel
```

and then hitting the `a` key on the function of interest.